[refactor] stats controller에서 로직 service로 분리
- stats에서 period, interval을 계산하는 로직 service로 분리
